Ajustes
- Se desacopla el service protocol
- Se hacen ajustes para separar las extensiones
