+++ BAZProject/MainTabController/Extensions/URLSessionExtension.swift
@@ -1,0 +1,16 @@
+//
+//  URLSessionExtension.swift
+//  BAZProject
+//
+//  Created by 989438 on 31/03/23.
+//
+
+import Foundation
+
+extension URLSessionDataTask: URLSessionDataTaskProtocol {}
+
+extension URLSession: URLSessionProtocol {
+    func performDataTask(with request: URLRequest, completionHandler: @escaping DataTaskResult) -> URLSessionDataTaskProtocol {
+        return dataTask(with: request, completionHandler: completionHandler) as URLSessionDataTaskProtocol
+    }
+}
