+++ BAZProject/MainTabController/Networking/EndPoint.swift
@@ -16,27 +16,3 @@ enum Endpoint {
     case topRated
     case upcoming
 }
-
-extension Endpoint: EndPointProtocol {
-    var string: String {
-        switch self {
-        case .trending:
-            return MainConstants.trendingEndPoint
-        case .nowPlaying:
-            return MainConstants.nowPlayingEndPoint
-        case .popular:
-            return MainConstants.popularEndPoint
-        case .topRated:
-            return MainConstants.topRatedEndPoint
-        case .upcoming:
-            return MainConstants.upcomingEndPoint
-        }
-    }
-    
-    var request: URLRequest {
-        let url: URL = URL(string: Endpoint.baseURL + string) ?? URL(fileURLWithPath: "")
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        return request
-    }
-}
