+++ BAZProject/MainTabController/Networking/ServiceAPI.swift
@@ -7,15 +7,7 @@
 
 import Foundation
 
-extension URLSessionDataTask: URLSessionDataTaskProtocol {}
-
-extension URLSession: URLSessionProtocol {
-    func performDataTask(with request: URLRequest, completionHandler: @escaping DataTaskResult) -> URLSessionDataTaskProtocol {
-        return dataTask(with: request, completionHandler: completionHandler) as URLSessionDataTaskProtocol
-    }
-}
-
-class ServiceAPI: Service {
+final class ServiceAPI: Service {
     let session: URLSessionProtocol
     
     init(session: URLSessionProtocol) {
