+++ BAZProject/MainTabController/Reusables/MainConstants.swift
@@ -18,22 +18,3 @@ enum MainConstants {
     static let upcomingEndPoint: String = "movie/upcoming?api_key=\(MainConstants.apiKey)&language=es&region=MX&page=1"
     
 }
-
-enum AlertSheetActions {
-    case trending
-    case nowPlaying
-    case popular
-    case topRated
-    case upcoming
-    
-    init?(rawValue: Int) {
-        switch rawValue {
-        case 0: self = .trending
-        case 1: self = .nowPlaying
-        case 2: self = .popular
-        case 3: self = .topRated
-        case 4: self = .upcoming
-        default: return nil
-        }
-    }
-}
