+++ BAZProject/MainTabController/TrendingModule/Router/TrendingBarRouter.swift
@@ -22,15 +22,9 @@ final class TrendingBarRouter: TrendingBarRouterProtocol {
         return view
     }
     
-    func presentSearchViewController(from view: TrendingBarViewControllerProtocol?) {
+    func presentNextViewController(from view: TrendingBarViewControllerProtocol?, to nextView: UIViewController) {
         guard let viewController = view as? UIViewController else { return }
-        let searchViewController  = SearchMoviesViewController()
-        viewController.navigationController?.pushViewController(searchViewController, animated: true)
-    }
-    
-    func presentMovieDetailController(from view: TrendingBarViewControllerProtocol?) {
-        guard let viewController = view as? UIViewController else { return }
-        let detailViewController  = MovieDetailViewController()
-        viewController.navigationController?.pushViewController(detailViewController, animated: true)
+        let nextViewController  = nextView
+        viewController.navigationController?.pushViewController(nextViewController, animated: true)
     }
 }
